+++ 3dEngine/src/utils/filter/gaussianblur/GaussianBlurFilter.cpp
@@ -13,7 +13,7 @@ namespace urchin
 		TextureFilter(textureFilterBuilder),
 		blurDirection(blurDirection),
 		blurSize(textureFilterBuilder->getBlurSize()),
-		nbTextureFetch(std::ceil(blurSize / 2.0f)),
+		nbTextureFetch(std::ceil((float)blurSize / 2.0f)),
 		textureSize((BlurDirection::VERTICAL==blurDirection) ? getTextureHeight() : getTextureWidth())
 	{ //See http://rastergrid.com/blog/2010/09/efficient-gaussian-blur-with-linear-sampling/
 
@@ -88,17 +88,17 @@ namespace urchin
 	{
 		std::vector<float> offsetsLinearSampling(nbTextureFetch);
 
-		int firstOffset = -static_cast<int>(std::floor(blurSize / 2.0f));
-		for(unsigned int i=0; i<nbTextureFetch; ++i)
+		int firstOffset = -(int)(std::floor((float)blurSize / 2.0f));
+		for(int i=0; i<nbTextureFetch; ++i)
 		{
-			if(i*2+1>=blurSize)
+			if(i * 2 + 1>=blurSize)
 			{
 				offsetsLinearSampling[i] = firstOffset+i*2;
 			}else
 			{
-				int offset1 = firstOffset+i*2;
+				int offset1 = firstOffset + i * 2;
 				int offset2 = offset1 + 1;
-				offsetsLinearSampling[i] = (offset1*weights[i*2] + offset2*weights[i*2+1]) / weightsLinearSampling[i];
+				offsetsLinearSampling[i] = ((float)offset1 * weights[i * 2] + (float)offset2 * weights[i * 2 + 1]) / weightsLinearSampling[i];
 			}
 
 			offsetsLinearSampling[i] /= static_cast<float>(textureSize);
